+++ python/cugraph/cugraph/community/__init__.py
@@ -22,6 +22,7 @@ from cugraph.community.spectral_clustering import (
     analyzeClustering_ratio_cut,
 )
 from cugraph.community.subgraph_extraction import subgraph
+from cugraph.community.induced_subgraph import induced_subgraph
 from cugraph.community.triangle_count import triangle_count
 from cugraph.community.ktruss_subgraph import ktruss_subgraph
 from cugraph.community.ktruss_subgraph import k_truss
